requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.1
attrs==25.4.0
Authlib==1.6.6
av==16.1.0
backoff==2.2.1
beautifulsoup4==4.14.3
browser-use==0.11.2
browser-use-sdk==2.0.12
bubus==1.5.6
cachetools==6.2.4
cdp-use==1.4.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
coloredlogs==15.0.1
cryptography==46.0.3
ctranslate2==4.6.3
Cython==3.2.4
distro==1.9.0
docstring_parser==0.17.0
easyocr==1.7.2
faster-whisper==1.2.1
filelock==3.20.3
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.1.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
google-genai==1.55.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
groq==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.1
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
inquirerpy==0.3.4
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy_loader==0.4
lxml==6.0.2
markdown-it-py==4.0.0
markdownify==1.2.2
MarkupSafe==3.0.3
mcp==1.25.0
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
mss==10.1.0
multidict==6.7.0
networkx==3.6.1
ninja==1.13.0
numpy==2.2.6
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.23.2
openai==2.15.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
outcome==1.3.0.post0
packaging==25.0
pfzy==0.3.4
pillow==12.1.0
playwright==1.57.0
portalocker==2.10.1
posthog==7.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
psutil==7.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pyclipper==1.4.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyee==13.0.0
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.10.1
PyMsgBox==2.0.1
pynput==1.8.1
pyobjc-core==12.1
pyobjc-framework-ApplicationServices==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-CoreText==12.1
pyobjc-framework-Quartz==12.1
pyotp==2.9.0
pyparsing==3.3.1
pypdf==6.5.0
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.21
pytweening==1.2.0
PyYAML==6.0.3
referencing==0.37.0
reportlab==4.4.7
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpa==1.50.0
rpds-py==0.30.0
rsa==4.9.1
rubicon-objc==0.5.3
scikit-image==0.26.0
scipy==1.16.3
screeninfo==0.8.1
selenium==4.39.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.3
soundfile==0.13.1
soupsieve==2.8.1
SpeechRecognition==3.14.5
sse-starlette==3.1.2
starlette==0.50.0
sympy==1.14.0
tagui==1.50.0
tenacity==9.1.2
tifffile==2025.12.20
tokenizers==0.22.2
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uuid7==0.1.0
uvicorn==0.40.0
wcwidth==0.2.14
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==15.0.1
wsproto==1.3.2
yarl==1.22.0
